level1 map finished + respawn when time runs out
