--- Assets/Timer.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 1bd0754c86b5b1943bd647406fe4790b
