--- Assets/playerControl.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9c6fe4d36d742154982cd48bf36bf19a
